seperate test for copy circuit
